refactor: reorganize GNN utilities and improve registry handling
