+++ mava/utils/make_env.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Tuple, Type
+from typing import Tuple, Type, TypeAlias
 
 import gymnasium
 import gymnasium as gym
@@ -63,8 +63,10 @@ from mava.wrappers import (
 from mava.wrappers.graph_wrapper import GraphWrapper
 from mava.wrappers.jaxmarl import MPEGraphWrapper
 
+registry_type: TypeAlias = dict[str, dict[str, Type]]
+
 # Registry mapping environment names to their generator and wrapper classes.
-_jumanji_registry: dict[str, dict[str, Type]] = {
+_jumanji_registry: registry_type = {
     "RobotWarehouse": {"generator": RwareRandomGenerator, "wrapper": RwareWrapper},
     "LevelBasedForaging": {"generator": LbfRandomGenerator, "wrapper": LbfWrapper},
     "Connector": {"generator": ConnectorRandomGenerator, "wrapper": ConnectorWrapper},
@@ -76,15 +78,15 @@ _jumanji_registry: dict[str, dict[str, Type]] = {
 }
 
 # Registry mapping environment names directly to the corresponding wrapper classes.
-_matrax_registry: dict[str, dict[str, Type]] = {"Matrax": {"wrapper": MatraxWrapper}}
-_jaxmarl_registry: dict[str, dict[str, Type]] = {
+_matrax_registry: registry_type = {"Matrax": {"wrapper": MatraxWrapper}}
+_jaxmarl_registry: registry_type = {
     "Smax": {"wrapper": SmaxWrapper},
     "MaBrax": {"wrapper": MabraxWrapper},
     "MPE": {"wrapper": MPEWrapper, "graph_wrapper": MPEGraphWrapper},
 }
-_gigastep_registry: dict[str, dict[str, Type]] = {"Gigastep": {"wrapper": GigastepWrapper}}
+_gigastep_registry: registry_type = {"Gigastep": {"wrapper": GigastepWrapper}}
 
-_gym_registry: dict[str, dict[str, Type]] = {
+_gym_registry: registry_type = {
     "RobotWarehouse": {"wrapper": UoeWrapper},
     "LevelBasedForaging": {"wrapper": UoeWrapper},
     "SMACLite": {"wrapper": SmacWrapper},
@@ -92,33 +94,24 @@ _gym_registry: dict[str, dict[str, Type]] = {
 
 
 def add_extra_wrappers(
-    train_env: MarlEnv, eval_env: MarlEnv, config: DictConfig
+    train_env: MarlEnv, eval_env: MarlEnv, config: DictConfig, registry: registry_type
 ) -> Tuple[MarlEnv, MarlEnv]:
+    """Wrappers that access and modify observations (like AgentIDWrapper) must come before
+    GraphWrapper to avoid special casing observation handling for both regular and graph
+    observations. For example, AgentIDWrapper adds agent IDs to observations, which should happen
+    before converting observation to GraphObservation."""
     # Disable the AgentID wrapper if the environment has implicit agent IDs.
     config.system.add_agent_id = config.system.add_agent_id & (~config.env.implicit_agent_id)
-
-    if is_gnn_based(config):
-        # Get the appropriate registry based on environment type
-        registry: dict[str, dict[str, Type]]
-        if config.env.env_name in _jumanji_registry:
-            registry = _jumanji_registry
-        elif config.env.env_name in _jaxmarl_registry:
-            registry = _jaxmarl_registry
-        elif config.env.env_name in _matrax_registry:
-            registry = _matrax_registry
-        elif config.env.env_name in _gigastep_registry:
-            registry = _gigastep_registry
-        else:
-            registry = _gym_registry
-
-        # Get the graph wrapper from registry or use default GraphWrapper
-        graph_wrapper = registry[config.env.env_name].get("graph_wrapper", GraphWrapper)
-        train_env = graph_wrapper(train_env)
-        eval_env = graph_wrapper(eval_env)
 
     if config.system.add_agent_id:
         train_env = AgentIDWrapper(train_env)
         eval_env = AgentIDWrapper(eval_env)
+
+    if is_gnn_based(config):
+        # Get the graph wrapper from registry or use default GraphWrapper
+        graph_wrapper = registry[config.env.env_name].get("graph_wrapper", GraphWrapper)
+        train_env = graph_wrapper(train_env)
+        eval_env = graph_wrapper(eval_env)
 
     train_env = AutoResetWrapper(train_env)
     train_env = RecordEpisodeMetrics(train_env)
@@ -154,7 +147,7 @@ def make_jumanji_env(config: DictConfig, add_global_state: bool = False) -> Tupl
     train_env = wrapper(train_env, add_global_state=add_global_state)
     eval_env = wrapper(eval_env, add_global_state=add_global_state)
 
-    train_env, eval_env = add_extra_wrappers(train_env, eval_env, config)
+    train_env, eval_env = add_extra_wrappers(train_env, eval_env, config, _jumanji_registry)
     return train_env, eval_env
 
 
@@ -189,7 +182,7 @@ def make_jaxmarl_env(config: DictConfig, add_global_state: bool = False) -> Tupl
         add_global_state,
     )
 
-    train_env, eval_env = add_extra_wrappers(train_env, eval_env, config)
+    train_env, eval_env = add_extra_wrappers(train_env, eval_env, config, _jaxmarl_registry)
 
     return train_env, eval_env
 
@@ -219,7 +212,7 @@ def make_matrax_env(config: DictConfig, add_global_state: bool = False) -> Tuple
     train_env = wrapper(train_env, add_global_state)
     eval_env = wrapper(eval_env, add_global_state)
 
-    train_env, eval_env = add_extra_wrappers(train_env, eval_env, config)
+    train_env, eval_env = add_extra_wrappers(train_env, eval_env, config, _matrax_registry)
     return train_env, eval_env
 
 
@@ -248,7 +241,7 @@ def make_gigastep_env(
     train_env: MarlEnv = wrapper(scenario.make(**kwargs), has_global_state=add_global_state)
     eval_env: MarlEnv = wrapper(scenario.make(**kwargs), has_global_state=add_global_state)
 
-    train_env, eval_env = add_extra_wrappers(train_env, eval_env, config)
+    train_env, eval_env = add_extra_wrappers(train_env, eval_env, config, _gigastep_registry)
     return train_env, eval_env
